Crate trampoline_sdk

Trampoline SDK

Welcome to Trampoline SDK documentation!

Trampoline is a Software Development Kit for creating full-stack decentralized applications on Nervos Network’s Common Knowledge Base blockchain.

Usage

Depends on trampoline-sdk in Cargo.toml:

[dependencies]
trampoline-sdk = { git = "https://github.com/Tempest-Protocol/trampoline", path = "sdk", branch = "develop" }

Please note that Trampoline is in early stages and under active development. The API can change drastically.

Here is an example of how to simulate deploying a cell on a CKB blockchain:

use trampoline_sdk::chain::{Chain,MockChain,CellInputs};
use trampoline_sdk::cell::Cell;
use std::collections::HashMap;

let unlockers = HashMap::new();
let inputs = CellInputs::Empty;

let mut chain = MockChain::default();
let cell = Cell::default();
let outpoint = chain.deploy_cell(&cell,unlockers, &inputs).expect("Failed to deploy cell");

Features

Trampoline is intended to be used in projects with a variety of compile targets and provides different features for each one:

Feature	Description

std	Standard set of features for non-contract development.

script	Reduced set of features for creating CKB on-chain scripts.

rpc	Support for interacting with CKB nodes and indexers through RPC.

Disabled features can be selectively enabled in Cargo.toml:

[dependencies]
trampoline-sdk = { git = "https://github.com/Tempest-Protocol/trampoline", path = "sdk", branch = "develop", features = ["rpc"] }

Modules

account

Types for handling signing on CKB transactions

address

Types for handling CKB addresses

bytes

Wrapper for different kind of Bytes objects

cell

Types for handling CKB cells

chain

Types and Traits for interacting with CKB blockchains

ckb_hash

Types and methods for handling Blake2b hashes

ckb_types

Types from the official CKB types lib

constants

Hashes and default configs for CKB

contract

Modules usable within on-chain scripts

precompiled

Precompiled contracts for creating standard and non-fungible tokens

query

Types for creating custom queries for CKB Indexer

rpc

Clients for interacting with CKB services

script

Types for handling CKB scripts

transaction

Types for handling and creating CKB transactions

Macros

impl_entity_unpack

impl_pack_for_fixed_byte_array

impl_pack_for_primitive

impl_primitive_reader_unpack