Struct trampoline_sdk::cell::Cell

pub struct Cell { /* private fields */ }

Cells are the primary state units in CKB and assets owned by users.

They must follow associated validation rules specified by scripts.

Implementations

impl Cell

pub fn with_data(data: impl Into<Bytes>) -> Self

Create a new Cell with a data field from bytes.

pub fn with_lock(script: impl Borrow<Script>) -> Self

Create a new Cell with a specific lock.

pub fn required_capacity(&self) -> CellResult<Capacity>

Returns the capacity required for the cell to hold itself

pub fn validate(&self) -> CellResult<Capacity>

Ensure the total cell size <= min required capacity and that the capacity in the cell >= min required capacity

pub fn lock_hash(&self) -> CellResult<H256>

Get the cell’s lock hash

pub fn type_hash(&self) -> CellResult<Option<H256>>

Get the cell’s type hash

pub fn type_script(&self) -> CellResult<Option<Script>>

Get the cell’s type script, if it has one

pub fn lock_script(&self) -> CellResult<Script>

Get the cell’s lock script

pub fn capacity(&self) -> Capacity

Get the cell’s capacity

pub fn data_size(&self) -> usize

Get the cell’s data size in bytes

pub fn outpoint(&self) -> Option<OutPoint>

Get the cell’s outpoint, if it was included in a transaction

pub fn data_hash(&self) -> H256

Get the cell’s data hash

pub fn data(&self) -> Bytes

Get the cell’s data as bytes

pub fn set_lock_script(&mut self, script: impl Into<Script>) -> CellResult<()>

Set the cell’s lock script

pub fn set_type_script(&mut self, script: impl Into<Script>) -> CellResult<()>

Set the cell’s type script

pub fn set_lock_args(&mut self, bytes: impl Into<Bytes>) -> CellResult<()>

Set the arguments for the cell’s lock script

pub fn set_type_args(&mut self, bytes: impl Into<Bytes>) -> CellResult<()>

Set the arguments for the cell’s type script

pub fn set_data(&mut self, data: impl Into<Bytes>) -> CellResult<()>

Set the cell’s data field

pub fn set_outpoint(&mut self, outp: OutPoint) -> CellResult<()>

Set the cell’s outpoint

pub fn set_capacity_ckbytes(&mut self, capacity: u64) -> CellResult<()>

Set the cell’s capacity in CKB (1 CKB equals to 10**8 shannons)

pub fn set_capacity_shannons(&mut self, capacity: u64) -> CellResult<()>

Set the cell’s capacity in Shannons (1 CKB equals to 10**8 shannons)

pub fn as_cell_dep(&self, _dep_type: DepType) -> CellResult<CellDep>

Build the cell as a CellDep to be used in transactions without being consumed