Struct trampoline_sdk::chain::mock_chain::MockChain

pub struct MockChain {
    pub cells: HashMap<OutPoint, CellOutputWithData>,
    pub outpoint_txs: HashMap<OutPoint, TransactionInfo>,
    pub headers: HashMap<Byte32, HeaderView>,
    pub epoches: HashMap<Byte32, EpochExt>,
    pub cells_by_data_hash: HashMap<Byte32, OutPoint>,
    pub cells_by_lock_hash: HashMap<Byte32, Vec<OutPoint>>,
    pub cells_by_type_hash: HashMap<Byte32, Vec<OutPoint>>,
    pub default_lock: Option<OutPoint>,
    pub debug: bool,
    /* private fields */
}

Simulated CKB chain for development and testing

Fields
cells: HashMap<OutPoint, CellOutputWithData>

Collection of all deployed cells

outpoint_txs: HashMap<OutPoint, TransactionInfo>

Collection of transaction information for each cell outpoint

headers: HashMap<Byte32, HeaderView>

Collection of all added block headers

epoches: HashMap<Byte32, EpochExt>

Collection of all epochs

cells_by_data_hash: HashMap<Byte32, OutPoint>

Collection of all deployed cells that have data

cells_by_lock_hash: HashMap<Byte32, Vec<OutPoint>>

Collection of all deployed cells by lock hash

cells_by_type_hash: HashMap<Byte32, Vec<OutPoint>>

Collection of all deployed cells by type hash

default_lock: Option<OutPoint>

Current default lock used by Mockchain methods

debug: bool

Debug flag

Implementations

impl MockChain

pub fn deploy_cell_with_data(&mut self, data: Bytes) -> OutPoint

Deploys a cell with given data

pub fn deploy_cell_output(
&mut self,
data: Bytes,
output: CellOutput
) -> OutPoint

Deploy a cell output to the mockchain

pub fn get_default_script_outpoint(&self) -> OutPoint

Get the outpoint for the current default script

pub fn deploy_random_cell_with_default_lock(
&mut self,
capacity: usize,
args: Option<Bytes>
) -> OutPoint

Deploy a cell with specified capacity using the default lock

pub fn insert_header(&mut self, header: HeaderView)

Add a block’s header to the mockchain

pub fn link_cell_with_block(
&mut self,
outp: OutPoint,
hash: Byte32,
tx_idx: usize
)

Link a cell with a particular block already in the chain

pub fn get_cell_by_data_hash(&self, data_hash: &Byte32) -> Option<OutPoint>

Get a deployed cell matching a data hash

pub fn create_cell(&mut self, cell: CellOutput, data: Bytes) -> OutPoint

Deploy a cell to the chain

pub fn create_cell_with_outpoint(
&mut self,
outp: OutPoint,
cell: CellOutput,
data: Bytes
)

Deploy a cell with an specific outpoint to the chain

pub fn get_cell(&self, out_point: &OutPoint) -> Option<CellOutputWithData>

Get a deployed cell from the chain

pub fn build_script_with_hash_type(
&self,
outp: &OutPoint,
typ: ScriptHashType,
args: Bytes
) -> Option<Script>

Build a Script from an outpoint and arguments

pub fn get_cells_by_lock_hash(&self, hash: Byte32) -> Option<Vec<OutPoint>>

Get all cells that match a lock hash

pub fn get_cells_by_type_hash(&self, hash: Byte32) -> Option<Vec<OutPoint>>

Get all cells that match a type hash

pub fn build_script(&self, outp: &OutPoint, args: Bytes) -> Option<Script>

Build a Script with HashType Data from outpoint and args

pub fn find_cell_dep_for_script(&self, script: &Script) -> CellDep

Get CellDep from a Script

pub fn complete_tx(&mut self, tx: TransactionView) -> TransactionView

pub fn build_resolved_tx(&self, tx: &TransactionView) -> ResolvedTransaction

Resolve a Transaction by dereferencing the pointers to cells

pub fn capture_debug(&self) -> bool

Returns true if debug is on

pub fn set_capture_debug(&mut self, capture_debug: bool)

Capture debug output, default value is false

pub fn captured_messages(&self) -> Vec<Message>

return captured messages

pub fn verify_tx_by_context(
&self,
tx: &TransactionView,
max_cycles: u64
) -> Result<Cycle, CKBError>

Verify the transaction by given context (Consensus, TxVerifyEnv) in CKB-VM always Please see below links for more details:

https://docs.rs/ckb-chain-spec/0.101.2/ckb_chain_spec/consensus/struct.Consensus.html

https://docs.rs/ckb-types/0.101.2/ckb_types/core/hardfork/struct.HardForkSwitch.html

https://docs.rs/ckb-script/0.101.2/ckb_script/struct.TxVerifyEnv.html

pub fn verify_tx(
&self,
tx: &TransactionView,
max_cycles: u64
) -> Result<Cycle, CKBError>

Verify the transaction in CKB-VM

This method use a default verify context with:

use HardForkSwitch to set rfc_0032 field to 200 (means enable VM selection feature after epoch 200)

use TxVerifyEnv to set currently transaction epoch number to 300

pub fn receive_tx(&mut self, tx: &TransactionView) -> Result<Byte32, CKBError>

Deploy each cell in a Transaction to the moockchain