Struct trampoline_sdk::chain::rpc_chain::provider::RpcProvider

pub struct RpcProvider { /* private fields */ }

Main data provider type for RpcChain

Implementations

impl RpcProvider

pub fn new(chain: RpcChain) -> Self

Creates a new RpcProvider from a RpcChain

pub fn get_tx(
&self,
hash: H256
) -> Result<Option<TransactionWithStatus>, ChainError>

Get a transaction by its hash, return it with its status

Transaction status may vary, it may be pending, proposed, commited or rejected

pub fn rollback(&self, previous_block: u64) -> Result<(), ChainError>

Truncate the Chain to a previous block height. Only available for devchains.

pub fn get_cell_with_data(
&self,
out_point: &OutPoint
) -> Result<(CellOutput, Bytes), TransactionDependencyError>

Gell a pair of CellOutput and Bytes from the chain

pub fn get_tip(&self) -> Option<HeaderView>

Get the last block’s header

pub fn genesis_block(&self) -> Result<BlockView, ChainError>

Get the chain’s genesis block

pub fn mine_once(&self) -> Result<H256, ChainError>

Mine a block with all pending transactions. Only available for devchains.

impl RpcProvider

pub fn resolve_tx(
&self,
tx: Transaction
) -> Result<ResolvedTransaction, ChainError>

Creates a ResolvedTransaction by dereferencing its outpoints

impl RpcProvider

pub fn get_consensus(&self) -> Consensus

Get Consensus object from chain