Trait trampoline_sdk::chain::Chain

pub trait Chain {
    type Inner: TransactionProvider;

    fn inner(&self) -> Self::Inner;
    fn deploy_cell(
        &mut self, 
        cell: &Cell, 
        unlockers: Unlockers, 
        inputs: &CellInputs
    ) -> ChainResult<OutPoint>;
    fn deploy_cells(
        &mut self, 
        cells: &[Cell], 
        unlockers: Unlockers, 
        inputs: &CellInputs
    ) -> ChainResult<Vec<OutPoint>>;
    fn set_default_lock(&mut self, lock: OutPoint) -> Result<(), ChainError>;
    fn generate_cell_with_default_lock(&self, lock_args: Bytes) -> Cell;

    fn verify_tx<T: Into<TransactionView> + Clone>(
        &self, 
        tx: T
    ) -> ChainResult<T> { ... }
    fn send_tx<T: Into<TransactionView> + Clone>(
        &self, 
        tx: T
    ) -> ChainResult<Byte32> { ... }
}

Universal API for interacting with CKB blockchains

This trait is implemented by [RpcChain] and [MockChain]. It provides a standard interface for verifying and sending transactions, deploying cells and specifying the default lock these methods should use.

Required Associated Types

type Inner: TransactionProvider

Inner type that provides transactions verification and sending methods

Required Methods

fn inner(&self) -> Self::Inner

Returns the inner Transaction Prpvider

fn deploy_cell(
&mut self,
cell: &Cell,
unlockers: Unlockers,
inputs: &CellInputs
) -> ChainResult<OutPoint>

Deploys a Cell to the chain

Cell Inputs must be provided, but if empty cells from the Unlocker address will be used.

fn deploy_cells(
&mut self,
cells: &[Cell],
unlockers: Unlockers,
inputs: &CellInputs
) -> ChainResult<Vec<OutPoint>>

Deploys multiples Cells to the chain

Cell Inputs must be provided, but if empty cells from the Unlocker address will be used.

fn set_default_lock(&mut self, lock: OutPoint) -> Result<(), ChainError>

Changes the chain default lock

fn generate_cell_with_default_lock(&self, lock_args: Bytes) -> Cell

Creates a default cell using the Chain’s default lock

Provided Methods

fn verify_tx<T: Into<TransactionView> + Clone>(&self, tx: T) -> ChainResult<T>

Verify if a transaction is valid given the current chain state

fn send_tx<T: Into<TransactionView> + Clone>(
&self,
tx: T
) -> ChainResult<Byte32>

Publish a transaction the the chain, returns its hash