Module trampoline_sdk::ckb_types

Types from the official CKB types lib

Modules

bytes

Provides abstractions for working with bytes.

constants

All Constants.

core

The essential rust types for CKB.

molecule

packed

Various newtypes of bytes::Bytes, and their associated types.

prelude

This module includes several traits.

utilities

Types utilities.

Macros

h160

A macro used to create a const H160 from a hexadecimal string or a trimmed hexadecimal string.

h256

A macro used to create a const H256 from a hexadecimal string or a trimmed hexadecimal string.

u256

Structs

H160

The 20-byte fixed-length binary data.

H256

The 32-byte fixed-length binary data.

U128

Fixed non-negative integer type.

U256

Fixed non-negative integer type.