Module trampoline_sdk::ckb_types::core::cell

TODO(doc): @quake

Structs

BlockCellProvider

TODO(doc): @quake

CellMeta

TODO(doc): @quake

CellMetaBuilder

TODO(doc): @quake

OverlayCellChecker

Overlay cell checker wrapper

OverlayCellProvider

TODO(doc): @quake

ResolvedTransaction

Transaction with resolved input cells.

TransactionsChecker

Cell checker for txs chain

TransactionsProvider

TODO(doc): @quake

Enums

CellStatus

TODO(doc): @quake

ResolvedDep

TODO(doc): @quake

Statics

SYSTEM_CELL

system cell memory map cache

Traits

CellChecker

Trait for check cell status

CellProvider

TODO(doc): @quake

HeaderChecker

TODO(doc): @quake

Functions

get_related_dep_out_points

Gather all cell dep out points and resolved dep group out points

resolve_transaction

Resolve all cell meta from db base on the transaction.

setup_system_cell_cache

TODO(doc): @quake

Type Definitions

SystemCellMap

type alias system cells map