Struct trampoline_sdk::ckb_types::core::TransactionView

pub struct TransactionView { /* private fields */ }

A readonly and immutable struct which includes Transaction and its associated hashes.

Notice

This struct is not implement the trait Default, use TransactionBuilder to construct it.

Implementations

impl TransactionView

pub fn new_advanced_builder() -> TransactionBuilder

Creates an empty advanced builder.

pub fn as_advanced_builder(&self) -> TransactionBuilder

Creates an advanced builder base on current data.

impl TransactionView

pub fn data(&self) -> Transaction

Gets a clone of packed::Transaction.

pub fn hash(&self) -> Byte32

Gets a clone of hash.

pub fn witness_hash(&self) -> Byte32

Gets a clone of witness_hash.

pub fn version(&self) -> u32

Gets raw.version.

pub fn cell_deps(&self) -> CellDepVec

Gets raw.cell_deps.

pub fn header_deps(&self) -> Byte32Vec

Gets raw.header_deps.

pub fn inputs(&self) -> CellInputVec

Gets raw.inputs.

pub fn outputs(&self) -> CellOutputVec

Gets raw.outputs.

pub fn outputs_data(&self) -> BytesVec

Gets raw.outputs_data.

pub fn witnesses(&self) -> BytesVec

Gets witnesses.

pub fn output(&self, idx: usize) -> Option<CellOutput>

Gets an output through its index.

pub fn output_with_data(&self, idx: usize) -> Option<(CellOutput, Bytes)>

Gets an output and its data through its index.

pub fn output_pts(&self) -> Vec<OutPoint, Global>

Gets out points for all outputs.

pub fn output_pts_iter(&self) -> impl Iterator<Item = OutPoint>

Creates an iterator from out points of all outputs.

pub fn input_pts_iter(&self) -> impl Iterator<Item = OutPoint>

Creates an iterator from out points of all inputs.

pub fn outputs_with_data_iter(
&self
) -> impl Iterator<Item = (CellOutput, Bytes)>

Creates an iterator from all outputs and their data.

pub fn cell_deps_iter(&self) -> impl Iterator<Item = CellDep>

Creates an iterator from raw.cell_deps.

pub fn header_deps_iter(&self) -> impl Iterator<Item = Byte32>

Creates an iterator from raw.header_deps.

pub fn fake_hash(self, hash: Byte32) -> TransactionView

Sets a fake transaction hash.

pub fn fake_witness_hash(self, witness_hash: Byte32) -> TransactionView

Sets a fake witness hash.

pub fn outputs_capacity(&self) -> Result<Capacity, Error>

Sums the capacities of all outputs.

pub fn is_cellbase(&self) -> bool

Checks whether the transaction is a cellbase.

pub fn proposal_short_id(&self) -> ProposalShortId

Creates a new ProposalShortId from the transaction hash.