Module trampoline_sdk::ckb_types::packed

Various newtypes of bytes::Bytes, and their associated types.

These newtypes are thread-safe and immutable binary data with a series of associated methods to read and convert themselves.

These newtypes are generated by Molecule from schemas.

References:

Using the Newtype Pattern to Implement External Traits on External Types

Using the Newtype Pattern for Type Safety and Abstraction

Structs

AddFilter

AddFilterBuilder

AddFilterReader

Address

AddressBuilder

AddressReader

AddressVec

AddressVecBuilder

AddressVecIterator

AddressVecReader

AddressVecReaderIterator

Alert

AlertBuilder

AlertReader

BeUint32

BeUint32Builder

BeUint32Reader

BeUint64

BeUint64Builder

BeUint64Reader

Block

BlockBuilder

BlockExt

BlockExtBuilder

BlockExtReader

BlockProposal

BlockProposalBuilder

BlockProposalReader

BlockReader

BlockTransactions

BlockTransactionsBuilder

BlockTransactionsReader

BlockV1

BlockV1Builder

BlockV1Reader

Bool

BoolBuilder

BoolOpt

BoolOptBuilder

BoolOptReader

BoolReader

Byte

Byte32

Byte32Builder

Byte32Opt

Byte32OptBuilder

Byte32OptReader

Byte32Reader

Byte32Vec

Byte32VecBuilder

Byte32VecIterator

Byte32VecReader

Byte32VecReaderIterator

ByteReader

Bytes

BytesBuilder

BytesIterator

BytesOpt

BytesOptBuilder

BytesOptReader

BytesReader

BytesVec

BytesVecBuilder

BytesVecIterator

BytesVecReader

BytesVecReaderIterator

CellDataEntry

CellDataEntryBuilder

CellDataEntryReader

CellDep

CellDepBuilder

CellDepReader

CellDepVec

CellDepVecBuilder

CellDepVecIterator

CellDepVecReader

CellDepVecReaderIterator

CellEntry

CellEntryBuilder

CellEntryReader

CellInput

CellInputBuilder

CellInputReader

CellInputVec

CellInputVecBuilder

CellInputVecIterator

CellInputVecReader

CellInputVecReaderIterator

CellOutput

CellOutputBuilder

CellOutputOpt

CellOutputOptBuilder

CellOutputOptReader

CellOutputReader

CellOutputVec

CellOutputVecBuilder

CellOutputVecIterator

CellOutputVecReader

CellOutputVecReaderIterator

CellbaseWitness

CellbaseWitnessBuilder

CellbaseWitnessReader

ClearFilter

ClearFilterBuilder

ClearFilterReader

CompactBlock

CompactBlockBuilder

CompactBlockReader

CompactBlockV1

CompactBlockV1Builder

CompactBlockV1Reader

DiscoveryMessage

DiscoveryMessageBuilder

DiscoveryMessageReader

DiscoveryPayload

DiscoveryPayloadBuilder

DiscoveryPayloadReader

EpochExt

EpochExtBuilder

EpochExtReader

FilteredBlock

FilteredBlockBuilder

FilteredBlockReader

GetBlockProposal

GetBlockProposalBuilder

GetBlockProposalReader

GetBlockTransactions

GetBlockTransactionsBuilder

GetBlockTransactionsReader

GetBlocks

GetBlocksBuilder

GetBlocksReader

GetHeaders

GetHeadersBuilder

GetHeadersReader

GetNodes

GetNodesBuilder

GetNodesReader

GetRelayTransactions

GetRelayTransactionsBuilder

GetRelayTransactionsReader

Header

HeaderBuilder

HeaderReader

HeaderVec

HeaderVecBuilder

HeaderVecIterator

HeaderVecReader

HeaderVecReaderIterator

HeaderView

HeaderViewBuilder

HeaderViewReader

Identify

IdentifyBuilder

IdentifyMessage

IdentifyMessageBuilder

IdentifyMessageReader

IdentifyReader

InIBD

InIBDBuilder

InIBDReader

IndexTransaction

IndexTransactionBuilder

IndexTransactionReader

IndexTransactionVec

IndexTransactionVecBuilder

IndexTransactionVecIterator

IndexTransactionVecReader

IndexTransactionVecReaderIterator

MerkleProof

MerkleProofBuilder

MerkleProofReader

Node

NodeBuilder

NodeReader

NodeVec

NodeVecBuilder

NodeVecIterator

NodeVecReader

NodeVecReaderIterator

Nodes

NodesBuilder

NodesReader

NumberHash

NumberHashBuilder

NumberHashReader

OutPoint

OutPointBuilder

OutPointReader

OutPointVec

OutPointVecBuilder

OutPointVecIterator

OutPointVecReader

OutPointVecReaderIterator

Ping

PingBuilder

PingMessage

PingMessageBuilder

PingMessageReader

PingPayload

PingPayloadBuilder

PingPayloadReader

PingReader

Pong

PongBuilder

PongReader

PortOpt

PortOptBuilder

PortOptReader

ProposalShortId

ProposalShortIdBuilder

ProposalShortIdReader

ProposalShortIdVec

ProposalShortIdVecBuilder

ProposalShortIdVecIterator

ProposalShortIdVecReader

ProposalShortIdVecReaderIterator

RawAlert

RawAlertBuilder

RawAlertReader

RawHeader

RawHeaderBuilder

RawHeaderReader

RawTransaction

RawTransactionBuilder

RawTransactionReader

RelayMessage

RelayMessageBuilder

RelayMessageReader

RelayTransaction

RelayTransactionBuilder

RelayTransactionHashes

RelayTransactionHashesBuilder

RelayTransactionHashesReader

RelayTransactionReader

RelayTransactionVec

RelayTransactionVecBuilder

RelayTransactionVecIterator

RelayTransactionVecReader

RelayTransactionVecReaderIterator

RelayTransactions

RelayTransactionsBuilder

RelayTransactionsReader

Script

ScriptBuilder

ScriptOpt

ScriptOptBuilder

ScriptOptReader

ScriptReader

SendBlock

SendBlockBuilder

SendBlockReader

SendHeaders

SendHeadersBuilder

SendHeadersReader

SetFilter

SetFilterBuilder

SetFilterReader

SyncMessage

SyncMessageBuilder

SyncMessageReader

Time

TimeBuilder

TimeReader

Transaction

TransactionBuilder

TransactionInfo

TransactionInfoBuilder

TransactionInfoReader

TransactionKey

TransactionKeyBuilder

TransactionKeyReader

TransactionReader

TransactionVec

TransactionVecBuilder

TransactionVecIterator

TransactionVecReader

TransactionVecReaderIterator

TransactionView

TransactionViewBuilder

TransactionViewReader

Uint16

Uint16Builder

Uint16Reader

Uint32

Uint32Builder

Uint32Reader

Uint32Vec

Uint32VecBuilder

Uint32VecIterator

Uint32VecReader

Uint32VecReaderIterator

Uint64

Uint64Builder

Uint64Reader

Uint64Vec

Uint64VecBuilder

Uint64VecIterator

Uint64VecReader

Uint64VecReaderIterator

Uint128

Uint128Builder

Uint128Reader

Uint256

Uint256Builder

Uint256Reader

UncleBlock

UncleBlockBuilder

UncleBlockReader

UncleBlockVec

UncleBlockVecBuilder

UncleBlockVecIterator

UncleBlockVecReader

UncleBlockVecReaderIterator

UncleBlockVecView

UncleBlockVecViewBuilder

UncleBlockVecViewReader

WitnessArgs

WitnessArgsBuilder

WitnessArgsReader

Enums

DiscoveryPayloadUnion

DiscoveryPayloadUnionReader

PingPayloadUnion

PingPayloadUnionReader

RelayMessageUnion

RelayMessageUnionReader

SyncMessageUnion

SyncMessageUnionReader