Module trampoline_sdk::ckb_types::prelude

This module includes several traits.

Few traits are re-exported from other crates, few are used as aliases and others are syntactic sugar.

Traits

Builder

Entity

FromSliceShouldBeOk

An alias of from_slice(..) to mark where we are really have confidence to do unwrap on the result of from_slice(..).

Pack

A syntactic sugar to convert a rust type into binary data.

PackVec

A syntactic sugar to convert a vector of binary data into one binary data.

Reader

ShouldBeOk

An alias of unwrap() to mark where we are really have confidence to do unwrap.

Unpack

A syntactic sugar to convert binary data into rust types.

Functions

hex_string