Trait trampoline_sdk::ckb_types::prelude::PackVec

pub trait PackVec<T, I>: IntoIterator<Item = I> where
    T: Entity,
    I: Entity, {
    fn pack(self) -> T;
}

A syntactic sugar to convert a vector of binary data into one binary data.

Required Methods

fn pack(self) -> T

Packs a vector of binary data into one binary data.