Module trampoline_sdk::ckb_types::utilities

Types utilities.

Structs

MergeByte32

TODO(doc): @quake

Constants

DIFF_TWO

The minimal difficulty that can be represented in the compact format.

Functions

compact_to_difficulty

Converts compact format of difficulty to the decoded difficulty.

compact_to_target

Converts compact format of difficulty to PoW target.

difficulty_to_compact

Converts difficulty into the compact format.

merkle_root

TODO(doc): @quake

target_to_compact

Converts PoW target into compact format of difficulty.

Type Definitions

CBMT

TODO(doc): @quake

MerkleProof

TODO(doc): @quake