Struct trampoline_sdk::transaction::CellMetaTransaction

pub struct CellMetaTransaction {
    pub tx: TransactionView,
    pub inputs: Vec<CellMeta>,
}

Type that combines a TX and input cells used by the generator

Fields
tx: TransactionView

Full transaction view

inputs: Vec<CellMeta>

Cells included in the transaction

Implementations

impl CellMetaTransaction

pub fn tx(self, tx: TransactionView) -> Self

Return a copy using the specified Transaction

pub fn with_inputs(self, inputs: Vec<CellMeta>) -> Self

Returns a copy using the specified inputs

pub fn as_advanced_builder(&self) -> TransactionBuilder

Returns a TransactionBuilder from the inner TX

pub fn cell_deps(&self) -> CellDepVec

Returns CellDeps from the inner TX

pub fn inputs(&self) -> CellInputVec

Returns CellInputs from the inner TX

pub fn outputs(&self) -> CellOutputVec

Returns CellOutputs from the inner TX

pub fn outputs_data(&self) -> BytesVec

Returns outputs data vector from the inner TX

pub fn witnesses(&self) -> BytesVec

Returns Witnesses from the inner TX

pub fn output(&self, idx: usize) -> Option<CellOutput>

Returns Output with specified index from the inner TX

pub fn output_with_data(&self, idx: usize) -> Option<CellOutputWithData>

Returns Output with data with specified index from the inner TX

pub fn output_pts(&self) -> Vec<OutPoint>

Returns Outpoints the inner TX

pub fn cell_deps_iter(&self) -> impl Iterator<Item = CellDep>

Returns CellDeps as an interator from the inner TX

pub fn output_pts_iter(&self) -> impl Iterator<Item = OutPoint>

Returns output Outpoints as an interator from the inner TX

pub fn input_pts_iter(&self) -> impl Iterator<Item = OutPoint>

Returns input outpoints as an interator from the inner TX

pub fn outputs_with_data_iter(&self) -> impl Iterator<Item = CellOutputWithData>

Returns outputs with data as an interator from the inner TX

pub fn outputs_capacity(&self) -> Result<Capacity, CapacityError>

Returns outputs total capacity from the inner TX

pub fn fake_hash(self, hash: Byte32) -> Self

Sets the hash of the inner transaction to an unchecked one

pub fn fake_witness_hash(self, witness_hash: Byte32) -> Self

Sets a fake witness hash.